1.10.10 Task: Code Cleanup (intellig project analysis)

+++ src/main/de/jClipCorn/gui/frames/addSeasonFrame/AddSeasonFrame.java
@@ -103,7 +103,7 @@ public class AddSeasonFrame extends JCCFrame implements UserDataProblemHandler, 
 		var spack = new SeasonDataPack
 				(
 						edTitle.getText(),
-						(int) spnYear.getValue(),
+						spnYear.getValue(),
 						edCvrControl.getResizedImageForStorage()
 				);
